requirements.txt
streamlit
pdfplumber
python-docx
pandas
openpyxl
python-pptx
transformers
sentence-transformers
chromadb
langchain
openai
